# operation-service/docs/swagger.yaml
basePath: /
definitions:
...
  model.PayloadCreateTransaction:
    description: PayloadCreateTransaction details.
    properties:
      grand_total:
        type: number
      order_id:
        type: string
      payment_method:
        allOf:
        - $ref: '#/definitions/model.PaymentMethod'
        enum:
        - cod
        - bank_transfer
    required:
    - grand_total
    - order_id
    - payment_method
    type: object
...
  model.PayloadUpdateTransaction:
    description: PayloadUpdateTransaction details.
    properties:
      grand_total:
        type: number
      payment_method:
        allOf:
        - $ref: '#/definitions/model.PaymentMethod'
        enum:
        - cod
        - bank_transfer
      payment_status:
        allOf:
        - $ref: '#/definitions/model.PaymentStatus'
        enum:
        - pending
        - completed
        - rejected
    type: object
  model.PaymentMethod:
    enum:
    - cod
    - bank_transfer
    type: string
    x-enum-varnames:
    - PaymentMethodCOD
    - PaymentMethodBankTransfer
  model.PaymentStatus:
    enum:
    - pending
    - completed
    - rejected
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusCompleted
    - PaymentStatusRejected
...
  model.ResponseTransaction:
    description: ResponseTransaction details.
    properties:
      created_at:
        type: string
      grand_total:
        type: number
      id:
        type: string
      order_id:
        type: string
      payment_method:
        $ref: '#/definitions/model.PaymentMethod'
      payment_status:
        $ref: '#/definitions/model.PaymentStatus'
      transaction_date:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Transaction:
    properties:
      created_at:
        type: string
      grand_total:
        type: number
      id:
        type: string
      order_id:
        type: string
      payment_method:
        $ref: '#/definitions/model.PaymentMethod'
      payment_status:
        $ref: '#/definitions/model.PaymentStatus'
      transaction_date:
        type: string
      updated_at:
        type: string
    type: object
  model.WebResponse:
    description: represents the standard API response format.
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: operation-service-84457363535.asia-southeast2.run.app
info:
  contact: {}
  description: This is the documentation of Operation Service API
  title: Operation Service API
  version: "1.0"
paths:
...
  /api/transactions:
    get:
      description: Find All Transactions
      produces:
      - application/json
      responses:
        "200":
          description: Transaction list
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Transaction'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: FindAll Transactions
      tags:
      - Transactions
    post:
      description: Add and register new Transaction
      parameters:
      - description: create transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PayloadCreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction created
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ResponseTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a Transaction
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      description: Delete Transaction By ID
      parameters:
      - description: transaction id
        example: '"67cdcb62a50a990a870d928f"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: DeleteById a Transaction
      tags:
      - Transactions
    get:
      description: Find Transaction By ID
      parameters:
      - description: transaction id
        example: '"67cdcb62a50a990a870d928f"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction found
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ResponseTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: FindById a Transaction
      tags:
      - Transactions
    put:
      description: Update Transaction By ID
      parameters:
      - description: transaction id
        example: '"67cdcb62a50a990a870d928f"'
        in: path
        name: id
        required: true
        type: string
      - description: update transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PayloadUpdateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ResponseTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a Transaction
      tags:
      - Transactions
  /api/transactions/approved/{id}:
    put:
      description: Update Status Payment to Completed By ID
      parameters:
      - description: transaction id
        example: '"67cdcb62a50a990a870d928f"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ResponseTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a Payment Status to Completed
      tags:
      - Transactions
  /api/transactions/reject/{id}:
    put:
      description: Update Status Payment to Reject By ID
      parameters:
      - description: transaction id
        example: '"67cdcb62a50a990a870d928f"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ResponseTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                code:
                  type: integer
                data: {}
                status:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a Payment Status to Reject
      tags:
      - Transactions
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"